updating Y - Range sum query problem

+++ sheet #3 array/Y. Range sum query.cpp
@@ -1,34 +1,33 @@
-
 #include <iostream>
 using namespace std;
  
  
  
- int rangeSum(int i, int j, int sum[]) {
+ long long rangeSum(long long i, long long j, long long sum[]) {
    if (i == 1)
       return sum[j-1];
    return sum[j-1] - sum[i - 2];
 }
  
 int main () {
-    int n,q;
+    long long n,q;
     cin>>n >> q;
     
     
-    int a[n], sum[n];
-    for (int i=0; i<n; i++) {
+    long long a[n], sum[n];
+    for (long long i=0; i<n; i++) {
         cin >> a[i];
     }
     sum[0] = a[0];
     // cout << sum[0] <<"------"<<endl;
-    for (int i = 1; i <= n; i++) {
+    for (long long i = 1; i <= n; i++) {
       sum[i] = a[i] + sum[i - 1];
     //   cout <<sum[i] << "------"<< endl;
     }
 
     //  cout<< rangeSum (1,6,sum);
-    for (int i=0; i<q; i++) {
-        int x,y;
+    for (long long i=0; i<q; i++) {
+        long long x,y;
         cin >> x >>y;
         cout<< rangeSum (x,y,sum) <<endl;
     }
